feat: Add exponential_base parameter to radial blur and depth map blur
This commit adds an "exponential_base" parameter to the "ProPostRadialBlur" and "ProPostDepthMapBlur" classes. This parameter allows for more control over the blurring effect by applying an exponential function to calculate the blur size. The default value is 1, and it can be adjusted between 1.0 and 8.0 with a step of 0.1.

The "generate_blurred_images" function in the "processing.py" file has been updated to use this new parameter when generating blurred images.